refactor:  renomeados e reorganizados componentes do perfil do usuário
