--- apps/frontend/src/app/(paginas)/(internas)/perfil/page.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-import Perfil from "@/src/views/perfil/Perfil";
-
-const Page: React.FC = () => {
-  return <Perfil />;
-};
-
-export default Page;
