+++ apps/frontend/src/app/(paginas)/(internas)/perfilusuario/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import PerfilUsuario from "@/src/views/perfilUsuario/PerfilUsuario";
+
+const Page: React.FC = () => {
+  return <PerfilUsuario />;
+};
+
+export default Page;
