+++ apps/frontend/src/app/(paginas)/trocarsenha/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import TrocaSenha from "@/src/views/perfil/TrocaSenha";
+import TrocaSenha from "@/src/views/perfilUsuario/TrocaSenha";
 
 const Page: React.FC = () => {
   return <TrocaSenha />;
